+++ spring/src/main/java/umc/spring/repository/ReviewRepository/ReviewRepositoryCustom.java
@@ -6,5 +6,5 @@ import umc.spring.domain.Review;
 import java.util.List;
 
 public interface ReviewRepositoryCustom {
-    Review dynamicQueryInsertReview(Long userId, String content, Long score, Long storeId);
+    Review dynamicQueryInsertReview(Long userId, String content, Float score, Long storeId);
 }
