+++ spring/src/main/java/umc/spring/service/ReviewService/ReviewQueryService.java
@@ -5,5 +5,5 @@ import umc.spring.domain.Review;
 import java.util.List;
 
 public interface ReviewQueryService {
-    Review addReview(Long userId, String content, Long score, Long storeId);
+    Review addReview(Long userId, String content, Float score, Long storeId);
 }
